refactor: add useRegisterForm.ts and use in RegisterPage.tsx
